requirements.txt
requests
pandas
python-telegram-bot
lxml